quantized elu: require observation
Summary:

ELU has a range of [-1, inf]. In the original PR which added
the quantized operator we decided to pass the quantization params
from the input.  However, it makes more sense to require observation
for this op.

This PR changes the API to require observation. Next PRs in this stack
will add the eager and graph mode handling.

Test Plan:

```
python test/test_quantization.py TestQuantizedOps.test_qelu
```

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 0d7a8005eb2ec001f4b96bb643b6d6ebeb136c48
Pull Request resolved: https://github.com/pytorch/pytorch/pull/40100